# contracts/openapi/footops-agent.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FootOps Agent API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/analyses": {
      "post": {
        "summary": "Create Analysis Workspace",
        "operationId": "create_analysis_workspace_api_v1_analyses_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisWorkspaceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisWorkspaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/stream": {
      "post": {
        "summary": "Stream Analysis Workspace",
        "operationId": "stream_analysis_workspace_api_v1_analyses_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisWorkspaceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Versioned analysis business events.",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/{workspace_id}": {
      "get": {
        "summary": "Get Analysis Workspace",
        "operationId": "get_analysis_workspace_api_v1_analyses__workspace_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisWorkspaceResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analyses/plan": {
      "post": {
        "summary": "Create Analysis Plan",
        "operationId": "create_analysis_plan_api_v1_analyses_plan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPlanResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/coverage-audit": {
      "post": {
        "summary": "Audit Player Data",
        "operationId": "audit_player_data_api_v1_data_coverage_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataCoverageAuditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/findings/player-role/review": {
      "post": {
        "summary": "Review Player Role Findings",
        "operationId": "review_player_role_findings_api_v1_findings_player_role_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerDataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerRoleFindingReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisPlan": {
        "properties": {
          "understanding": {
            "$ref": "#/components/schemas/QuestionUnderstanding"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/AnalysisPlanStep"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Steps"
          },
          "clarifying_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Clarifying Questions"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Assumptions"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Limitations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "understanding",
          "steps"
        ],
        "title": "AnalysisPlan",
        "description": "Structured question understanding and analysis plan."
      },
      "AnalysisPlanRequest": {
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1,
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AnalysisPlanRequest"
      },
      "AnalysisPlanResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "const": "planned",
            "title": "Status",
            "default": "planned"
          },
          "mode": {
            "type": "string",
            "enum": [
              "mock",
              "deepseek"
            ],
            "title": "Mode"
          },
          "provider": {
            "type": "string",
            "enum": [
              "none",
              "deepseek"
            ],
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "model_called": {
            "type": "boolean",
            "title": "Model Called"
          },
          "data_retrieved": {
            "type": "boolean",
            "title": "Data Retrieved",
            "default": false
          },
          "real_conclusions_generated": {
            "type": "boolean",
            "title": "Real Conclusions Generated",
            "default": false
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "plan": {
            "$ref": "#/components/schemas/AnalysisPlan"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "mode",
          "provider",
          "model",
          "model_called",
          "message",
          "plan"
        ],
        "title": "AnalysisPlanResponse"
      },
...
      "AnalysisWorkspace": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "status": {
            "$ref": "#/components/schemas/WorkspaceStatus",
            "default": "created"
          },
          "request": {
            "$ref": "#/components/schemas/AnalysisRequestArtifact"
          },
          "coverage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CoverageAuditArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "metrics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerRoleMetricArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingArtifact"
            },
            "type": "array",
            "title": "Findings"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceSetArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence_review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvidenceReviewArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "tactics_board": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TacticsBoardArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request"
        ],
        "title": "AnalysisWorkspace"
      },
      "AnalysisWorkspaceCreateRequest": {
        "properties": {
          "competition_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Competition Id"
          },
          "season_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Season Id"
          },
          "player": {
            "type": "string",
            "minLength": 1,
            "title": "Player"
          },
          "requested_window": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 3.0,
            "title": "Requested Window",
            "default": 5
          },
          "question": {
            "type": "string",
            "minLength": 1,
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "competition_id",
          "season_id",
          "player",
          "question"
        ],
        "title": "AnalysisWorkspaceCreateRequest"
      },
      "AnalysisWorkspaceResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ready",
            "title": "Status",
            "default": "ready"
          },
          "provider": {
            "type": "string",
            "const": "statsbomb-open-data",
            "title": "Provider",
            "default": "statsbomb-open-data"
          },
          "model_called": {
            "type": "boolean",
            "const": false,
            "title": "Model Called",
            "default": false
          },
          "data_retrieved": {
            "type": "boolean",
            "const": true,
            "title": "Data Retrieved",
            "default": true
          },
          "findings_generated": {
            "type": "boolean",
            "const": true,
            "title": "Findings Generated",
            "default": true
          },
          "evidence_reviewed": {
            "type": "boolean",
            "const": true,
            "title": "Evidence Reviewed",
            "default": true
          },
          "tactics_board_generated": {
            "type": "boolean",
            "const": true,
            "title": "Tactics Board Generated",
            "default": true
          },
          "real_conclusions_generated": {
            "type": "boolean",
            "const": false,
            "title": "Real Conclusions Generated",
            "default": false
          },
          "workspace": {
            "$ref": "#/components/schemas/AnalysisWorkspace"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workspace"
        ],
        "title": "AnalysisWorkspaceResponse"
      },
...
      "DataCoverageAuditResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "audited",
            "title": "Status",
            "default": "audited"
          },
          "provider": {
            "type": "string",
            "const": "statsbomb-open-data",
            "title": "Provider",
            "default": "statsbomb-open-data"
          },
          "model_called": {
            "type": "boolean",
            "const": false,
            "title": "Model Called",
            "default": false
          },
          "data_retrieved": {
            "type": "boolean",
            "const": true,
            "title": "Data Retrieved",
            "default": true
          },
          "real_conclusions_generated": {
            "type": "boolean",
            "const": false,
            "title": "Real Conclusions Generated",
            "default": false
          },
          "audit": {
            "$ref": "#/components/schemas/CoverageAuditArtifact"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "audit"
        ],
        "title": "DataCoverageAuditResponse"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "const": "error",
            "title": "Status",
            "default": "error"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerDataRequest": {
        "properties": {
          "competition_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Competition Id"
          },
          "season_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Season Id"
          },
          "player": {
            "type": "string",
            "minLength": 1,
            "title": "Player"
          },
          "requested_window": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 3.0,
            "title": "Requested Window",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "competition_id",
          "season_id",
          "player"
        ],
        "title": "PlayerDataRequest"
      },
...
      "PlayerRoleFindingReviewResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "reviewed",
            "title": "Status",
            "default": "reviewed"
          },
          "provider": {
            "type": "string",
            "const": "statsbomb-open-data",
            "title": "Provider",
            "default": "statsbomb-open-data"
          },
          "model_called": {
            "type": "boolean",
            "const": false,
            "title": "Model Called",
            "default": false
          },
          "data_retrieved": {
            "type": "boolean",
            "const": true,
            "title": "Data Retrieved",
            "default": true
          },
          "findings_generated": {
            "type": "boolean",
            "const": true,
            "title": "Findings Generated",
            "default": true
          },
          "evidence_reviewed": {
            "type": "boolean",
            "const": true,
            "title": "Evidence Reviewed",
            "default": true
          },
          "real_conclusions_generated": {
            "type": "boolean",
            "const": false,
            "title": "Real Conclusions Generated",
            "default": false
          },
          "audit": {
            "$ref": "#/components/schemas/CoverageAuditArtifact"
          },
          "metrics": {
            "$ref": "#/components/schemas/PlayerRoleMetricArtifact"
          },
          "findings": {
            "$ref": "#/components/schemas/FindingSetArtifact"
          },
          "evidence": {
            "$ref": "#/components/schemas/EvidenceSetArtifact"
          },
          "evidence_review": {
            "$ref": "#/components/schemas/EvidenceReviewArtifact"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "audit",
          "metrics",
          "findings",
          "evidence",
          "evidence_review"
        ],
        "title": "PlayerRoleFindingReviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}